requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
bcrypt==3.1.6
billiard==3.6.0.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.2.1
django-celery-results==1.0.4
django-rest-swagger==2.2.0
django-tagging==0.4.6
djangorestframework==3.9.3
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.15.0
entrypoints==0.3
flake8==3.7.7
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
kombu==4.5.0
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
meld3==1.0.2
openapi-codec==1.3.2
paramiko==2.4.2
psycopg2==2.8.1
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyNaCl==1.3.0
pytz==2019.1
PyYAML==3.13
redis==3.2.1
requests==2.21.0
ruamel.yaml==0.15.94
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
supervisor==4.0.2
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
websocket-client==0.56.0
